requirements.txt
Django==1.6.5
Unipath==1.0
dj-database-url==0.3.0
dj-static==0.0.5
gunicorn==18.0
python-decouple==2.2
psycopg2==2.5.3
pytz
south==0.8.4
mock==1.0.1
django-localflavor==1.0
python-social-auth==0.2.1
django-widget-tweaks==1.3
requests==2.5.1
Pillow==2.7.0